+++ connectors/src/connectors/slack/index.ts
@@ -629,24 +629,3 @@ export async function retrieveSlackChannelsTitles(
 
   return new Ok(titles);
 }
-
-export async function maybeLaunchSlackSyncWorkflowForChannelId(
-  connectorId: number,
-  slackChannelId: string
-) {
-  const channelId = await SlackChannel.findOne({
-    attributes: ["id"],
-    where: {
-      connectorId,
-      slackChannelId,
-    },
-  });
-
-  if (!channelId) {
-    throw new Error(`Slack channel ${slackChannelId} does not exist in DB.`);
-  }
-
-  return launchSlackSyncWorkflow(connectorId.toString(), null, [
-    slackChannelId,
-  ]);
-}
